Trimming down commonutils in preparation to build on more distros (#576)
