+++ src/common/commonutils/CMakeLists.txt
@@ -3,8 +3,6 @@
 
 project(commonutils)
 
-find_package(RapidJSON REQUIRED)
-
 add_library(commonutils STATIC 
     CommandUtils.c
     ConfigUtils.c
@@ -16,8 +14,7 @@ add_library(commonutils STATIC
     SocketUtils.c
     SshUtils.c
     UrlUtils.c
-    UserUtils.c
-    CommonUtils.cpp)
+    UserUtils.c)
 
 target_compile_options(commonutils PRIVATE -Wno-psabi -Wall -Wextra -Wunused -Werror -Wformat -Wformat-security -Wno-nonnull -Wno-unused-result -Wunused-const-variable=2 -Wunused-macros)
 
@@ -29,4 +26,5 @@ target_include_directories(commonutils
 
 target_link_libraries(commonutils PRIVATE 
     logging 
-    parsonlib)
+    parsonlib
+    m)
