+++ src/common/commonutils/CommonUtils.h
@@ -112,9 +112,6 @@ char* FormatAllocateString(const char* format, ...);
 size_t HashString(const char* source);
 char* HashCommand(const char* source, void* log);
 
-bool IsValidClientName(const char* name);
-bool IsValidMimObjectPayload(const char* payload, const int payloadSizeBytes, void* log);
-
 bool ParseHttpProxyData(const char* proxyData, char** hostAddress, int* port, char**username, char** password, void* log);
 
 char* GetOsName(void* log);
