+++ src/common/commonutils/Internal.h
@@ -35,17 +35,4 @@
 
 #define PLAIN_STATUS_FROM_ERRNO(a) ((0 == a) ? "passed" : "failed")
 
-#ifdef __cplusplus
-
-#include <algorithm>
-#include <cstdio>
-#include <cstring>
-#include <string>
-#include <regex>
-#include <rapidjson/document.h>
-#include <rapidjson/schema.h>
-#include <rapidjson/stringbuffer.h>
-
-#endif //__cplusplus
-
 #endif // INTERNAL_H
