+++ src/modules/ztsi/src/lib/Ztsi.h
@@ -10,6 +10,8 @@
 #define ZTSI_LOGFILE "/var/log/osconfig_ztsi.log"
 #define ZTSI_ROLLEDLOGFILE "/var/log/osconfig_ztsi.bak"
 
+bool IsValidClientName(const char* name);
+
 class ZtsiLog
 {
 public:
